updated printed parts to reduce led holders to 1 in the middle instead of 2. Also adjusted it to allow greater tolerance between diffusers.
Adjusted controller arduino code to reflect the new number of leds.

Updated the game code so that the lights will turn off when program exits.

Fixed bug where process would continue after game exited due to thread loop.

+++ Tankontroller/Scenes/GameOverScene.cs
@@ -28,7 +28,7 @@ namespace Tankontroller.Scenes
             int x = (screenWidth - width) / 2;
             int y = (screenHeight - height) / 2;
             mRectangle = new Rectangle(0, 0, screenWidth, screenHeight);
-            mSecondsLeft = DGS.Instance.GetInt("SECONDS_TO_DISPLAY_GAMEOVER_SCREEN");
+            mSecondsLeft = DGS.Instance.GetFloat("SECONDS_TO_DISPLAY_GAMEOVER_SCREEN");
             game.ReplaceCurrentMusicInstance("Music/Music_start", true);
             mPlayers = pPlayers;
             mWinner = pWinner;
